chore: Bitcoin heights and depths made uint32 (#181)
Resolves #179 

This is an impactful change that brings the BTC height types that we use
more in sync with the expected types for BTC block heights, similar to
those that are used by other libraries. It is part of the code
improvement efforts that we are doing in preparation of the next
testnet.

Given that it touches many files, I believe that this change should
happen sooner than later so that it can be appropriately tested.

+++ cmd/babylond/cmd/genhelpers/set_btc_delegations_test.go
@@ -67,8 +67,8 @@ func FuzzCmdSetBtcDels(f *testing.F) {
 		slashingPkScript, err := txscript.PayToAddrScript(slashingAddress)
 		require.NoError(t, err)
 
-		startHeight := datagen.RandomInt(r, 100) + 1
-		endHeight := datagen.RandomInt(r, 1000) + startHeight + btcctypes.DefaultParams().CheckpointFinalizationTimeout + 1
+		startHeight := uint32(datagen.RandomInt(r, 100)) + 1
+		endHeight := uint32(datagen.RandomInt(r, 1000)) + startHeight + btcctypes.DefaultParams().CheckpointFinalizationTimeout + 1
 		slashingRate := sdkmath.LegacyNewDecWithPrec(int64(datagen.RandomInt(r, 41)+10), 2)
 		slashingChangeLockTime := uint16(101)
